+++ app/Http/Controllers/StepAutocompleteController.php
@@ -7,11 +7,16 @@ use Issue\Http\Requests;
 use Issue\Http\Controllers\Controller;
 use Issue\StepAutocomplete;
 use Issue\Http\Requests\StepAutocompleteRequest;
+use Gate;
 
 class StepAutocompleteController extends Controller
 {
     public function index()
     {
+        if (Gate::denies('list-step-autocomplete')) {
+            abort(403);
+        }
+
         $steps = StepAutocomplete::all();
 
         return view('admin.backend.step-autocomplete.list', ['steps' => $steps]);
@@ -19,6 +24,10 @@ class StepAutocompleteController extends Controller
 
     public function store(StepAutocompleteRequest $request)
     {
+        if (Gate::denies('store-step-autocomplete')) {
+            abort(403);
+        }
+
         $step = new StepAutocomplete;
 
         $step->name = $request->input('name');
@@ -29,6 +38,10 @@ class StepAutocompleteController extends Controller
 
     public function destroy($stepautocomplete)
     {
+        if (Gate::denies('delete-step-autocomplete')) {
+            abort(403);
+        }
+
         $stepautocomplete->delete();
 
         return redirect()->action('StepAutocompleteController@index');
